requirements.txt
pillow>=9.0.0
qrcode>=7.3.1
opencv-python>=4.5.5.64
pyzbar>=0.1.9
numpy>=1.22.0
pyperclip>=1.8.2